Fix drag drop issues
Collections need to be mutable to allow drag/drop to change their position in the list, otherwise we get a COMException. Children being null on the item can also throw a COMException.

Additionally we're using drag events to alter the drop target from Inside to Before, if the target row's model isn't a category.

The RowsPresenter.ChildIndexChanged event is also being used to update the backing Index property, since this isn't a dependency property on TreeDataGridRow.